Rename old molecules module (#414)
* Add es steps to trajectory and fix struc frames

* Rename validation task document model

* Rename task state enum

* Rename old molecules module
